commit to fix heroku

+++ controllers/admin.js
@@ -1,3 +1,7 @@
+const mongoose = require('mongoose');
+
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator');
 
 const Product = require('../models/product');
@@ -15,12 +19,28 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: []
+    });
+  }
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
+    console.log(errors.array());
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/admin/add-product',
@@ -32,11 +52,15 @@ exports.postAddProduct = (req, res, next) => {
         price: price,
         description: description
       },
-    errorMessage: errors.array()[0].msg,
-    validationErrors: errors.array()
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array()
     });
   }
+
+  const imageUrl = image.path;
+
   const product = new Product({
+    // _id: new mongoose.Types.ObjectId('5badf72403fd8b5be0366e81'),
     title: title,
     price: price,
     description: description,
@@ -51,8 +75,23 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect('/admin/products');
     })
     .catch(err => {
+      // return res.status(500).render('admin/edit-product', {
+      //   pageTitle: 'Add Product',
+      //   path: '/admin/add-product',
+      //   editing: false,
+      //   hasError: true,
+      //   product: {
+      //     title: title,
+      //     imageUrl: imageUrl,
+      //     price: price,
+      //     description: description
+      //   },
+      //   errorMessage: 'Database operation failed, please try again.',
+      //   validationErrors: []
+      // });
+      // res.redirect('/500');
       const error = new Error(err);
-      err.httpStatusCode = 500;
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
@@ -80,7 +119,7 @@ exports.getEditProduct = (req, res, next) => {
     })
     .catch(err => {
       const error = new Error(err);
-      err.httpStatusCode = 500;
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
@@ -89,8 +128,9 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
+
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
@@ -101,13 +141,12 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
       },
-    errorMessage: errors.array()[0].msg,
-    validationErrors: errors.array()
+      errorMessage: errors.array()[0].msg,
+      validationErrors: errors.array()
     });
   }
 
@@ -119,22 +158,24 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
-      return product.save()
-      .then(result => {
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
+      return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
       });
     })
     .catch(err => {
       const error = new Error(err);
-      err.httpStatusCode = 500;
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find({userId: req.user._id})
+  Product.find({ userId: req.user._id })
     // .select('title price -_id')
     // .populate('userId', 'name')
     .then(products => {
@@ -147,21 +188,26 @@ exports.getProducts = (req, res, next) => {
     })
     .catch(err => {
       const error = new Error(err);
-      err.httpStatusCode = 500;
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })  
     .then(() => {
       console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
+      res.status(200).json({message: 'Success!'});
     })
     .catch(err => {
-      const error = new Error(err);
-      err.httpStatusCode = 500;
-      return next(error);
+      res.status(500).json({message: 'Deleting product failed.'})
     });
 };
